observeCafeを雑に分割
